mysite/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
dill==0.3.6
distlib==0.3.6
Django==4.2.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-multiupload==0.6.1
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.2
filelock==3.12.0
flake8==6.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.5.0
platformdirs==3.5.1
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sentry-sdk==1.25.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
wrapt==1.15.0
